device: build graphics allocator 2.0
errors:
I hwservicemanager: Tried to start android.hardware.graphics.allocator@2.0::IAllocator/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.
E init    : Control message: Could not find 'android.hardware.graphics.allocator@2.0::IAllocator/default' for ctl.interface_start from pid: 571 (/system/bin/hwservicemanager)